feat(public-project): add API to fetch project members by ID and update dependencies
