+++ apps/public/actions/upwithcrowd/public-project/actions.ts
@@ -1,8 +1,8 @@
 "use server";
 
-import {getUpwithcrowd} from "@/utils/client";
-import {GetApiPublicProjectProjectListData} from "@ayasofyazilim/upwithcrowd-saas/UPWCService";
+import type {GetApiPublicProjectProjectListData} from "@ayasofyazilim/upwithcrowd-saas/UPWCService";
 import {structuredError, structuredResponse} from "@repo/utils/api";
+import {getUpwithcrowd} from "@/utils/client";
 
 export async function getPublicProjectsApi(data?: GetApiPublicProjectProjectListData) {
   try {
@@ -18,6 +18,16 @@ export async function getPublicProjectDetailByIdApi(id: string) {
   try {
     const client = await getUpwithcrowd();
     const dataResponse = await client.publicProject.getApiPublicProjectProjectDetailById({id});
+    return structuredResponse(dataResponse);
+  } catch (error) {
+    return structuredError(error);
+  }
+}
+
+export async function getPublicProjectByIdMembersApi(id: string) {
+  try {
+    const client = await getUpwithcrowd();
+    const dataResponse = await client.publicProject.getApiPublicProjectByIdMembers({id});
     return structuredResponse(dataResponse);
   } catch (error) {
     return structuredError(error);
